# probo-api-spec.json
{
  "swagger": "2.0",
  "info": {
    "title": "Probo Reseller API",
    "description": "Reseller API",
    "version": "1.0.0"
  },
  "host": "api.proboprints.com",
  "schemes": [
    "https"
  ],
  "paths": {
    "/order": {
      "x-amf-description": "With this endpoint a reseller can place orders at Probo and receive order information.",
      "post": {
        "description": "Place an order at Probo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "type": "string"
          },
          {
            "x-amf-mediaType": "application/json",
            "in": "body",
            "name": "generated",
            "schema": {
              "$ref": "#/definitions/type"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "x-amf-mediaType": "application/json",
            "schema": {
              "example": {
                "status": "order_queued",
                "message": "Order queued",
                "order": {
                  "id": "1112",
                  "reference": "Reseller API test order"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or invalid client application credentials"
          },
          "500": {
            "description": "Bad response from authorization server, or WSDL SOAP Fault error"
          }
        }
      }
    },
    "/order/status": {
      "x-amf-description": "With this endpoint a reseller can request the status of orders placed at Probo.",
      "post": {
        "description": "Request status of an order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "type": "string"
          },
          {
            "x-amf-mediaType": "application/json",
            "in": "body",
            "name": "generated",
            "schema": {
              "$ref": "#/definitions/type_1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "x-amf-mediaType": "application/json",
            "schema": {
              "example": {
                "status": "ok",
                "message": "Status retrieved",
                "orders": [
                  {
                    "id": "order-5234231",
                    "reference": "test order",
                    "supplier_order_number": 201900001,
                    "status_code": "shipment_delivered",
                    "delivery_date": "2020-08-14",
                    "shipping_method_code": "dhl-letterbox",
                    "products": [
                      {
                        "id": 1,
                        "status_code": "shipment_delivered"
                      }
                    ],
                    "packages": [
                      {
                        "id": 1,
                        "status_code": "shipment_delivered",
                        "tracking_url": "https://dhlparcel.nl/en/private/receiving/follow-your-shipment?tt=JJD00009005735170002249&amp;pc=1181GG"
                      }
                    ]
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized or invalid client application credentials"
          },
          "500": {
            "description": "Bad response from authorization server, or WSDL SOAP Fault error"
          }
        }
      }
    },
    "/order/cancel": {
      "x-amf-description": "With this endpoint a reseller can cancel multiple order placed at Probo.",
      "post": {
        "description": "Cancel order",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "type": "string"
          },
          {
            "x-amf-mediaType": "application/json",
            "in": "body",
            "name": "generated",
            "schema": {
              "$ref": "#/definitions/type_2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "x-amf-mediaType": "application/json",
            "schema": {
              "example": {
                "status": "cancelled",
                "message": "Order cancelled",
                "id": "order-5234231"
              }
            }
          },
          "400": {
            "description": "",
            "x-amf-mediaType": "application/json",
            "schema": {
              "x-amf-examples": {
                "cannotcancel": {
                  "status": "error",
                  "message": "Could not cancel order, because order status is: production",
                  "id": "order-5234231"
                },
                "failed": {
                  "status": "error",
                  "message": "Failed to cancel order, please contact Probo",
                  "id": "order-5234231"
                },
                "notexists": {
                  "status": "error",
                  "message": "Order does not exist",
                  "id": "order-5234231"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or invalid client application credentials"
          },
          "500": {
            "description": "Bad response from authorization server, or WSDL SOAP Fault error"
          }
        }
      }
    },
...
    "/products/configure": {
      "x-amf-description": "Configure Probo product(s)",
      "post": {
        "description": "Configure Probo product(s)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "type": "string"
          },
          {
            "x-amf-mediaType": "application/json",
            "in": "body",
            "name": "generated",
            "schema": {
              "$ref": "#/definitions/type_8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "x-amf-mediaType": "application/json",
            "schema": {
              "example": {
                "value": {
                  "products": [
                    {
                      "api_code": "clearview",
                      "can_order": true,
                      "available_options": [],
                      "selected_options": [
                        {
                          "type_code": "width",
                          "name": "Breite",
                          "description": null,
                          "value": "100",
                          "code": null,
                          "default_value": null,
                          "min_value": "0.00",
                          "max_value": "99999.00",
                          "step_size": null,
                          "scale": 1,
                          "reversible": true
                        },
                        {
                          "type_code": "height",
                          "name": "Höhe",
                          "description": null,
                          "value": "100",
                          "code": null,
                          "default_value": null,
                          "min_value": "0.00",
                          "max_value": "99999.00",
                          "step_size": null,
                          "scale": 1,
                          "reversible": true
                        },
                        {
                          "type_code": "amount",
                          "name": null,
                          "description": null,
                          "value": "100",
                          "code": null,
                          "external_id": null,
                          "default_value": null,
                          "min_value": null,
                          "max_value": null,
                          "step_size": null,
                          "scale": null,
                          "reversible": null
                        },
                        {
                          "type_code": "radio",
                          "name": "Eine Schicht Vollfarbe",
                          "description": "Der Druck besteht aus einer Schicht Vollfarbe Tinte.",
                          "value": null,
                          "code": null,
                          "default_value": null,
                          "min_value": null,
                          "max_value": null,
                          "step_size": null,
                          "scale": null,
                          "reversible": null
                        },
                        {
                          "type_code": "radio",
                          "name": "Außenseite",
                          "description": "Die Folie wird an die Außenseite (des Fensters) geklebt.",
                          "value": null,
                          "code": null,
                          "default_value": null,
                          "min_value": null,
                          "max_value": null,
                          "step_size": null,
                          "scale": null,
                          "reversible": null
                        },
                        {
                          "type_code": "radio",
                          "name": "Ohne Laminat",
                          "description": "Die Folie wird nicht laminiert.",
                          "value": null,
                          "code": null,
                          "default_value": null,
                          "min_value": null,
                          "max_value": null,
                          "step_size": null,
                          "scale": null,
                          "reversible": null
                        },
                        {
                          "type_code": "radio",
                          "name": "Schneiden",
                          "description": "Die Folie wird auf das gewünschte Format zugeschnitten.",
                          "api_code": null,
                          "default_value": null,
                          "min_value": null,
                          "max_value": null,
                          "step_size": null,
                          "scale": null,
                          "reversible": null
                        }
                      ],
                      "amount": 100,
                      "width": 100,
                      "height": 100,
                      "length": 0,
                      "deliveries:": [
                        {
                          "delivery_date": "2019-12-07T00:00:00+0000",
                          "production_hours": 24,
                          "shipping_methods": [
                            {
                              "name": "Probo-Lieferung",
                              "code": "onbezorgd-standard",
                              "description": "Speziell für große Pakete mit einer Verpackungsgröße ab 180 cm. Für diese Pakete erhalten Sie zwischen 6:00 und 10:00 Uhr einen Track & Trace-Link in einer E-Mail. Über diesen Link können Sie sehen, in welchem Zeitfenster von zwei Stunden das Paket zugestellt wird.",
                              "prices": {
                                "purchase_price": 0,
                                "puchase_price_incl_vat": 0.0,
                                "sales_price": 0,
                                "sales_price_incl_vat": 0.0
                              }
                            }
                          ],
                          "prices_total": {
                            "purchase_price": 0,
                            "puchase_price_incl_vat": 0.0,
                            "sales_price": 0,
                            "sales_price_incl_vat": 0.0
                          },
                          "prices_per_product": {
                            "purchase_price": 0,
                            "puchase_price_incl_vat": 0.0,
                            "sales_price": 0,
                            "sales_price_incl_vat": 0.0
                          }
                        }
                      ]
                    }
                  ],
                  "calculation_id": 82606
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or invalid client application credentials"
          },
          "500": {
            "description": "Bad response from authorization server, or WSDL SOAP Fault error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "type": {
      "x-amf-examples": {
        "minimal": {
          "value": {
            "error_email_addresses": [
              "customer@test.com"
            ],
            "callback_urls": [
              "http://test.com/status"
            ],
            "id": "1112",
            "reference": "Reseller API test order",
            "contact_email": "customer@test.com",
            "language": "en",
            "deliveries": [
              {
                "address": {
                  "company_name": "Probo Sign",
                  "first_name": "John",
                  "last_name": "Doe",
                  "street": "Fortuinweg",
                  "house_number": "17",
                  "addition": "A",
                  "postal_code": "9101PE",
                  "city": "Dokkum",
                  "country": "NL",
                  "phone": "0031880223200",
                  "email": "test@test.nl"
                },
                "delivery_date": "2020-08-12",
                "shipping_method_code": "onbezorgd-standard"
              }
            ],
            "products": [
              {
                "code": "beachflag",
                "options": [
                  {
                    "code": "amount",
                    "value": 1
                  },
                  {
                    "code": "straight"
                  },
                  {
                    "code": "large"
                  },
                  {
                    "code": "left"
                  },
                  {
                    "code": "no-beachpole"
                  }
                ],
                "files": [
                  {
                    "uri": "http://test.com/printfile.pdf",
                    "pages": [
                      {
                        "page": 1
                      }
                    ]
                  }
                ]
              }
            ]
          }
        },
        "mingled-product-types": {
          "value": {
            "error_email_addresses": [
              "customer@test.com"
            ],
            "callback_urls": [
              "http://test.com/status"
            ],
            "id": "1112",
            "reference": "Reseller API test order",
            "contact_email": "customer@test.com",
            "language": "en",
            "deliveries": [
              {
                "address": {
                  "company_name": "Probo Sign",
                  "first_name": "John",
                  "last_name": "Doe",
                  "street": "Fortuinweg",
                  "house_number": "17",
                  "addition": "A",
                  "postal_code": "9101PE",
                  "city": "Dokkum",
                  "country": "NL",
                  "phone": "0031880223200",
                  "email": "test@test.nl"
                },
                "delivery_date": "2020-08-12",
                "shipping_method_code": "onbezorgd-standard"
              }
            ],
            "products": [
              {
                "code": "2748",
                "reference": "reference",
                "options": [
                  {
                    "code": "60883",
                    "value": 10
                  }
                ]
              },
              {
                "customer_code": "myBeachflag",
                "options": [
                  {
                    "code": "1234",
                    "value": 5
                  }
                ]
              }
            ]
          }
        }
      },
      "type": "object",
      "required": [
        "id",
        "reference",
        "deliveries",
        "products"
      ],
      "properties": {
        "error_email_addresses": {
          "description": "Array of email addresses which will receive status updates regarding the processing of the order.",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "callback_urls": {
          "description": "Array of url's which will be called with status updates regarding the processing of the order.",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "order_type": {
          "description": "Set this parameter to \"test\" for testing an order. When a test order is accepted, the order will be visible in your webshop account, but will not be taken into production. Default is \"production\"",
          "enum": [
            "production",
            "test"
          ],
          "type": "string"
        },
        "project": {
          "description": "Set this boolean to true if this is a project order.",
          "type": "boolean"
        },
        "id": {
          "description": "Reseller's identification of the order. Must be unique for each order **(can be re-used when order is cancelled!)**. May only contain the following characters: a-zA-Z0-9_- ",
          "type": "string",
          "pattern": "^[\\w-]+$",
          "minLength": 1
        },
        "reference": {
          "description": "Reseller's reference of the order. Does not have to be unique.",
          "type": "string",
          "minLength": 1
        },
        "contact_email": {
          "description": "Email of the reseller's contact person on behalf the order is placed for. A default contact email address can be set in the Reseller's webshop account settings.",
          "type": "string",
          "minLength": 5
        },
        "language": {
          "description": "The language set in the the Reseller's account is used as default, specify a different language here if necessary. Use ISO 639-1 standard. (not yet supported)",
          "type": "string",
          "minLength": 2,
          "maxLength": 2
        },
        "deliveries": {
          "type": "array",
          "minItems": 1,
          "items": {
            "x-amf-union": [
              {
                "$ref": "#/definitions/DeliveryDateAndShippingMethodCode"
              },
              {
                "$ref": "#/definitions/DeliveryDateAndShippingMethodPreset"
              },
              {
                "$ref": "#/definitions/DeliveryDatePresetAndShippingMethodCode"
              },
              {
                "$ref": "#/definitions/DeliveryDatePresetAndShippingMethodPreset"
              },
              {
                "$ref": "#/definitions/ShippingDateAndShippingMethodCode"
              },
              {
                "$ref": "#/definitions/ShippingDateAndShippingMethodPreset"
              }
            ]
          }
        },
        "products": {
          "type": "array",
          "minItems": 1,
          "items": {
            "x-amf-union": [
              {
                "$ref": "#/definitions/Code"
              },
              {
                "$ref": "#/definitions/CustomerCode"
              }
            ]
          }
        },
        "project_id": {
          "description": "Resellers project group what the order relates to.",
          "type": "integer"
        }
      }
    },
...
  }
}